# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whisper-wrap engine API",
    "description": "HTTP/WebSocket/SSE API for the whisper-wrap transcription engine.",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "3.0.0-beta.1"
  },
  "paths": {
...
    "/ask": {
      "post": {
        "tags": [
          "transcription"
        ],
        "operationId": "ask",
        "parameters": [
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "description": "Optional per-call model override (llm-provider-abstraction). Absent\nselects the active provider's default model.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "log",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "description": "Either a JSON body `{\"text\": \"…\"}` for a text question, or raw/multipart audio (`audio/*`, `application/octet-stream`, or `multipart/form-data` with a `file` part) to transcribe-then-ask.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Answer. When `stream=true` the response is a `text/event-stream` (SSE) of incremental `data:` chunks terminated by a final event; otherwise a single JSON answer object."
          },
          "400": {
            "description": "Empty or malformed body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Content-Type or media format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Transcription or LLM failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No ASR model loaded (audio input) or no LLM provider configured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/config/ai": {
      "get": {
        "tags": [
          "ai-config"
        ],
        "summary": "`GET /config/ai` — masked read of the active config.",
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "The masked, read-safe AI provider config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ai-config"
        ],
        "summary": "`PUT /config/ai` — validate, save, swap the live client, return the masked\nview. Invalid provider -> 400. Empty `apiKey` keeps the stored key.",
        "operationId": "put_config",
        "requestBody": {
          "description": "Partial AI provider config update (camelCase). Unknown keys are ignored; the runtime accepts a lenient JSON object.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated masked config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiConfigView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid config body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/config/ai/models": {
      "get": {
        "tags": [
          "ai-config"
        ],
        "summary": "`GET /config/ai/models` — list provider models using the SUBMITTED\nprovider/base-url/key. Any fetch failure -> 200 with `models: []` + a\nnon-null `error` (never a 5xx).",
        "operationId": "list_models",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "baseUrl",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available models for the given provider/baseUrl/apiKey, or an error descriptor embedded in the JSON."
          }
        }
      }
    },
    "/config/ai/test": {
      "post": {
        "tags": [
          "ai-config"
        ],
        "summary": "`POST /config/ai/test` — build a transient client from the submitted body,\ndo one minimal non-streaming `ask`, report `{ ok, error }`. Never persists.",
        "operationId": "test_config",
        "requestBody": {
          "description": "Provider credentials/settings to test-connect against.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connectivity test result `{ok, …}` (failures are reported in the body, not as a non-200 status)."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiConfigUpdate": {
        "type": "object",
        "description": "Update body for `PUT /config/ai`. camelCase on the wire.",
        "required": [
          "provider"
        ],
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "systemPrompt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AiConfigView": {
        "type": "object",
        "description": "The masked, read-safe view of the active config. Wire shape is camelCase.",
        "required": [
          "provider",
          "baseUrl",
          "model",
          "keySet",
          "keyHint",
          "systemPromptSet"
        ],
        "properties": {
          "baseUrl": {
            "type": "string"
          },
          "keyHint": {
            "type": "string"
          },
          "keySet": {
            "type": "boolean"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "systemPromptSet": {
            "type": "boolean"
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "The `{ \"detail\": string }` body every [`ApiError`] serializes to. `ApiError`\nitself is private and only implements `IntoResponse`, so it cannot derive\n`ToSchema`; this is the single reusable schema every fallible operation\nreferences for its non-200 error responses (rather than re-inlining the\nshape per handler).",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error description.",
            "example": "Unsupported Content-Type: text/plain"
          }
        }
      },
...
      "TestBody": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "engine_token_bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "engine_token_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "engine_token"
      }
    }
  },
  "security": [
    {
      "engine_token_bearer": []
    },
    {
      "engine_token_cookie": []
    }
  ]
}